UMUTIMA BACKEND/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai
Pillow==10.1.0
requests==2.31.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
